requirements.txt
requests~=2.31.0
Pillow~=10.2.0
python-docx~=1.1.0
tqdm~=4.66.2